fixed some bugs
moved ReceiptTemplate model to common

tweaked Receipt Model

fixed template.success_rate update issue
